requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
duckduckgo-search==5.3.1b1
openai==1.35.10
httpx
tavily-python
